Lower required Go version to 1.21
Starting in Go 1.21, the `go` directive sets the minimum version of Go required to use a module. This
meant that a project that had `go 1.21` in it's go.mod file could not use this module.
Now that is resolved. Go 1.21 is still a supported version so I see any problems with this. We could
also set it back to 1.18, as that is what https://github.com/bobg/mingo reports as the minimum Go version
we need. But I prefer 1.21 for now in case I want to introduce use of https://pkg.go.dev/testing#Testing
in the future.

https://go.dev/doc/modules/gomod-ref#go