feat: Preserve <snap:ignore> markers in DiffJSON
Adds the ability to preserve <snap:ignore> markers in JSON diffs when
updating the snapshot. Before this, any <snap:ignore> markers would be
replaced with the value returned from the function under test. This was
annoying, but somewhat easy to work around with git.

This implementation has some limitations. It is not perfect by any
means. It requires the snapshot diff of the JSON to have each field and
value be on it's own line. Like so:

```
		snap.Snap(t, `{
  "Value": "foo",
  "Time": "<snap:ignore>",
  "Vendor": "bar"
}`).DiffJSON(&req, "  ")
```

In my experience, I have kept fields on there own distinct line, so this
is an improvement that works for the common case. A more robust
implementation in the future may involve using JSON decoding to handle
JSON tokens better.